+++ main_test.go
@@ -451,14 +451,14 @@ func TestEntryPoint(t *testing.T) {
 		}()
 		time.Sleep(1 * time.Second)
 
-		t.Run("403 - without headers and collections", func(t *testing.T) {
+		t.Run("200 - even without headers", func(t *testing.T) {
 			gock.Flush()
 			gock.New("http://localhost:3002/users/").
 				Get("/users/").
 				Reply(200)
 			resp, err := http.DefaultClient.Get("http://localhost:3003/users/")
 			require.Equal(t, nil, err)
-			require.Equal(t, http.StatusForbidden, resp.StatusCode)
+			require.Equal(t, http.StatusOK, resp.StatusCode)
 		})
 		t.Run("200 - integration passed", func(t *testing.T) {
 			gock.Flush()
@@ -561,11 +561,92 @@ func TestEntryPoint(t *testing.T) {
 			Get("/users/").
 			Reply(200)
 		req, err := http.NewRequest("GET", "http://localhost:3034/users/", nil)
+		req.Header.Set("miausergroups", "group1")
 		req.Header.Set("miauserid", "filter_test")
 		client1 := &http.Client{}
 		resp, err := client1.Do(req)
 		require.Equal(t, nil, err)
 		require.Equal(t, http.StatusOK, resp.StatusCode)
+	})
+
+	t.Run("403 - integration not passed with query generation and without user authenticated", func(t *testing.T) {
+		shutdown := make(chan os.Signal, 1)
+
+		defer gock.Off()
+		gock.EnableNetworking()
+		gock.NetworkingFilter(func(r *http.Request) bool {
+			if r.URL.Path == "/documentation/json" {
+				return false
+			}
+			if r.URL.Path == "/users/" && r.URL.Host == "localhost:3035" {
+				return false
+			}
+			return true
+		})
+
+		gock.New("http://localhost:3035").
+			Get("/documentation/json").
+			Reply(200).
+			File("./mocks/simplifiedMockWithRowFiltering.json")
+
+		unsetBaseEnvs := setEnvs([]env{
+			{name: "HTTP_PORT", value: "3036"},
+			{name: "TARGET_SERVICE_HOST", value: "localhost:3035"},
+			{name: "TARGET_SERVICE_OAS_PATH", value: "/documentation/json"},
+			{name: "OPA_MODULES_DIRECTORY", value: "./mocks/rego-policies"},
+		})
+		mongoHost := os.Getenv("MONGO_HOST_CI")
+		if mongoHost == "" {
+			mongoHost = testutils.LocalhostMongoDB
+			t.Logf("Connection to localhost MongoDB, on CI env this is a problem!")
+		}
+		randomizedDBNamePart := testutils.GetRandomName(10)
+		mongoDBName := fmt.Sprintf("test-%s", randomizedDBNamePart)
+
+		unsetOtherEnvs := setEnvs([]env{
+			{name: "MONGODB_URL", value: fmt.Sprintf("mongodb://%s/%s", mongoHost, mongoDBName)},
+			{name: "BINDINGS_COLLECTION_NAME", value: "bindings"},
+			{name: "ROLES_COLLECTION_NAME", value: "roles"},
+		})
+
+		clientOpts := options.Client().ApplyURI(fmt.Sprintf("mongodb://%s", mongoHost))
+		client, err := mongo.Connect(context.Background(), clientOpts)
+		if err != nil {
+			fmt.Printf("error connecting to MongoDB: %s", err.Error())
+		}
+
+		ctx, cancelFn := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancelFn()
+		if err = client.Ping(ctx, readpref.Primary()); err != nil {
+			fmt.Printf("error verifying MongoDB connection: %s", err.Error())
+		}
+		mongoClient := MongoClient{
+			client:   client,
+			roles:    client.Database(mongoDBName).Collection("roles"),
+			bindings: client.Database(mongoDBName).Collection("bindings"),
+		}
+		defer mongoClient.client.Disconnect(ctx)
+
+		PopulateDbForTesting(t, ctx, &mongoClient)
+
+		go func() {
+			entrypoint(shutdown)
+		}()
+		defer func() {
+			unsetBaseEnvs()
+			unsetOtherEnvs()
+			shutdown <- syscall.SIGTERM
+		}()
+		time.Sleep(1 * time.Second)
+		gock.Flush()
+		gock.New("http://localhost:3035/users/").
+			Get("/users/").
+			Reply(200)
+		req, err := http.NewRequest("GET", "http://localhost:3036/users/", nil)
+		client1 := &http.Client{}
+		resp, err := client1.Do(req)
+		require.Equal(t, nil, err)
+		require.Equal(t, http.StatusForbidden, resp.StatusCode)
 	})
 }